# front-land/openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/messages/conversation": {
      "get": {
        "tags": [
          "Messages"
        ],
        "operationId": "GetConversation",
        "parameters": [
          {
            "name": "userId1",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId2",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationMessagesDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationMessagesDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationMessagesDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/user/{userId}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "operationId": "GetUserConversations",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/upload": {
      "post": {
        "tags": [
          "Messages"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/send": {
      "post": {
        "tags": [
          "Messages"
        ],
        "operationId": "SendMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/mark-read/{messageId}": {
      "put": {
        "tags": [
          "Messages"
        ],
        "operationId": "MarkMessageAsRead",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/unread-count/{userId}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "operationId": "GetUnreadCount",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/files/{filename}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/archive": {
      "post": {
        "tags": [
          "Messages"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/unarchive": {
      "post": {
        "tags": [
          "Messages"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/mute": {
      "post": {
        "tags": [
          "Messages"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/unmute": {
      "post": {
        "tags": [
          "Messages"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/block": {
      "post": {
        "tags": [
          "Messages"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/unblock": {
      "post": {
        "tags": [
          "Messages"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/delete-conversation": {
      "delete": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "otherUserId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/search": {
      "get": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/report": {
      "post": {
        "tags": [
          "Messages"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportAbuseRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportAbuseRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReportAbuseRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/notification/mark-as-read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "operationId": "MarkAsRead",
        "parameters": [
          {
            "name": "notificationId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatActionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "otherUserId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ConversationDto": {
        "type": "object",
        "properties": {
          "otherUserId": {
            "type": "integer",
            "format": "int32"
          },
          "otherUserName": {
            "type": "string",
            "nullable": true
          },
          "otherUserProfilePicture": {
            "type": "string",
            "nullable": true
          },
          "lastMessage": {
            "$ref": "#/components/schemas/MessageDto"
          },
          "unreadCount": {
            "type": "integer",
            "format": "int32"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isMuted": {
            "type": "boolean"
          },
          "isBlocked": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ConversationMessagesDto": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MessageDto": {
        "type": "object",
        "properties": {
          "messageId": {
            "type": "integer",
            "format": "int32"
          },
          "senderId": {
            "type": "integer",
            "format": "int32"
          },
          "receiverId": {
            "type": "integer",
            "format": "int32"
          },
          "messageText": {
            "type": "string",
            "nullable": true
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "isRead": {
            "type": "boolean"
          },
          "isSuperLike": {
            "type": "boolean"
          },
          "senderName": {
            "type": "string",
            "nullable": true
          },
          "receiverName": {
            "type": "string",
            "nullable": true
          },
          "senderProfilePicture": {
            "type": "string",
            "nullable": true
          },
          "receiverProfilePicture": {
            "type": "string",
            "nullable": true
          },
          "fileUrl": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "fileType": {
            "type": "string",
            "nullable": true
          },
          "isPriority": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ReportAbuseRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "reportedUserId": {
            "type": "integer",
            "format": "int32"
          },
          "messageId": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SendMessageInputDto": {
        "type": "object",
        "properties": {
          "receiverId": {
            "type": "integer",
            "format": "int32"
          },
          "messageText": {
            "type": "string",
            "nullable": true
          },
          "isSuperLike": {
            "type": "boolean"
          },
          "fileUrl": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "fileType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Please enter JWT with Bearer into field (e.g. 'Bearer {token}')",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}